test: fix reset
