+++ tests/setup.ts
@@ -1,6 +1,9 @@
-import { fillWindowEnv } from './utils';
+/* eslint-disable no-console, import/prefer-default-export */
+import util from 'util';
+import type { DOMWindow } from 'jsdom';
 
-/* eslint-disable no-console */
+// import { fillWindowEnv } from './utils';
+
 const React = require('react');
 
 // eslint-disable-next-line no-console
@@ -20,6 +23,42 @@ console.error = (...args) => {
     originConsoleErr(...args);
   }
 };
+
+type Writeable<T> = { -readonly [P in keyof T]: T[P] };
+
+export function fillWindowEnv(window: Window | DOMWindow) {
+  const win = window as Writeable<Window> & typeof globalThis;
+
+  win.resizeTo = (width, height) => {
+    win.innerWidth = width || win.innerWidth;
+    win.innerHeight = height || win.innerHeight;
+    win.dispatchEvent(new Event('resize'));
+  };
+  win.scrollTo = () => {};
+  // ref: https://github.com/ant-design/ant-design/issues/18774
+  if (!win.matchMedia) {
+    Object.defineProperty(win, 'matchMedia', {
+      writable: true,
+      configurable: true,
+      value: jest.fn((query) => ({
+        matches: query.includes('max-width'),
+        addListener: jest.fn(),
+        removeListener: jest.fn(),
+      })),
+    });
+  }
+
+  // Fix css-animation or rc-motion deps on these
+  // https://github.com/react-component/motion/blob/9c04ef1a210a4f3246c9becba6e33ea945e00669/src/util/motion.ts#L27-L35
+  // https://github.com/yiminghe/css-animation/blob/a5986d73fd7dfce75665337f39b91483d63a4c8c/src/Event.js#L44
+  win.AnimationEvent = win.AnimationEvent || win.Event;
+  win.TransitionEvent = win.TransitionEvent || win.Event;
+
+  // ref: https://jestjs.io/docs/manual-mocks#mocking-methods-which-are-not-implemented-in-jsdom
+  // ref: https://github.com/jsdom/jsdom/issues/2524
+  Object.defineProperty(win, 'TextEncoder', { writable: true, value: util.TextEncoder });
+  Object.defineProperty(win, 'TextDecoder', { writable: true, value: util.TextDecoder });
+}
 
 /* eslint-disable global-require */
 if (typeof window !== 'undefined') {
